ajustando o di
