+++ app/src/main/AndroidManifest.xml
@@ -6,7 +6,7 @@
 
     <application
         android:allowBackup="true"
-        android:name=".di.MyApp"
+        android:name=".common.di.MyApp"
         android:icon="@mipmap/ic_launcher"
         android:label="@string/app_name"
         android:networkSecurityConfig="@xml/network_security_config"
@@ -15,7 +15,7 @@
         android:theme="@style/AppTheme"
         tools:ignore="AllowBackup,GoogleAppIndexingWarning">
         <activity
-            android:name=".presentation.view.activity.MainActivity"
+            android:name=".users.presentation.view.activity.MainActivity"
             android:exported="true">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
