--- app/src/main/java/com/picpay/desafio/android/common/Mapper.kt
@@ -1,5 +1,0 @@
-package com.picpay.desafio.android.common
-
-interface Mapper<S, T> {
-    suspend fun map(source: S): T
-}
