+++ app/src/main/java/com/picpay/desafio/android/common/mapper/Mapper.kt
@@ -1,0 +1,5 @@
+package com.picpay.desafio.android.common.mapper
+
+interface Mapper<S, T> {
+    suspend fun map(source: S): T
+}
