--- app/src/main/java/com/picpay/desafio/android/data/api/PicPayService.kt
@@ -1,11 +1,0 @@
-package com.picpay.desafio.android.data.api
-
-import com.picpay.desafio.android.data.models.User
-import retrofit2.http.GET
-
-
-interface PicPayService {
-
-    @GET("users")
-    suspend fun getUsers(): List<User>
-}
