--- app/src/main/java/com/picpay/desafio/android/data/datasource/UsersDataSource.kt
@@ -1,7 +1,0 @@
-package com.picpay.desafio.android.data.datasource
-
-import com.picpay.desafio.android.data.models.User
-
-interface UsersDataSource {
-    suspend fun getUsers(): List<User>
-}
