--- app/src/main/java/com/picpay/desafio/android/data/mapper/UsersListMapper.kt
@@ -1,8 +1,0 @@
-package com.picpay.desafio.android.data.mapper
-
-import com.picpay.desafio.android.data.models.User
-import com.picpay.desafio.android.domain.models.UsersDomain
-
-interface UsersListMapper {
-    suspend fun map(source: User): UsersDomain
-}
