--- app/src/main/java/com/picpay/desafio/android/data/mapper/UsersListMapperImpl.kt
@@ -1,16 +1,0 @@
-package com.picpay.desafio.android.data.mapper
-
-import com.picpay.desafio.android.common.Mapper
-import com.picpay.desafio.android.data.models.User
-import com.picpay.desafio.android.domain.models.UsersDomain
-
-class UsersListMapperImpl : Mapper<User, UsersDomain>, UsersListMapper {
-    override suspend fun map(source: User): UsersDomain {
-        return UsersDomain(
-            id = source.id,
-            name = source.name,
-            username = source.username,
-            img = source.img
-        )
-    }
-}
