--- app/src/main/java/com/picpay/desafio/android/data/models/User.kt
@@ -1,14 +1,0 @@
-package com.picpay.desafio.android.data.models
-
-import com.google.gson.annotations.SerializedName
-
-data class User(
-    @SerializedName("img")
-    val img: String?,
-    @SerializedName("name")
-    val name: String?,
-    @SerializedName("id")
-    val id: String?,
-    @SerializedName("username")
-    val username: String?
-)
