--- app/src/main/java/com/picpay/desafio/android/di/MyApp.kt
@@ -1,21 +1,0 @@
-package com.picpay.desafio.android.di
-
-import android.app.Application
-import org.koin.android.ext.koin.androidContext
-import org.koin.core.context.startKoin
-import org.koin.dsl.KoinAppDeclaration
-
-class MyApp : Application() {
-
-    override fun onCreate() {
-        super.onCreate()
-        startKoin(appDeclaration = KoinAppDeclarationProvider.get(this))
-    }
-
-    object KoinAppDeclarationProvider {
-        fun get(application: Application): KoinAppDeclaration = {
-            androidContext(application)
-            UsersModule().load()
-        }
-    }
-}
