--- app/src/main/java/com/picpay/desafio/android/domain/models/UsersDomain.kt
@@ -1,8 +1,0 @@
-package com.picpay.desafio.android.domain.models
-
-data class UsersDomain(
-    val img: String?,
-    val name: String?,
-    val id: String?,
-    val username: String?
-)
