--- app/src/main/java/com/picpay/desafio/android/domain/repository/UsersRepository.kt
@@ -1,7 +1,0 @@
-package com.picpay.desafio.android.domain.repository
-
-import com.picpay.desafio.android.domain.models.UsersDomain
-
-interface UsersRepository {
-    suspend fun getUsers(): List<UsersDomain>
-}
