--- app/src/main/java/com/picpay/desafio/android/presentation/usecase/UserUseCase.kt
@@ -1,7 +1,0 @@
-package com.picpay.desafio.android.presentation.usecase
-
-import com.picpay.desafio.android.domain.models.UsersDomain
-
-interface UserUseCase {
-    suspend operator fun invoke(): List<UsersDomain>
-}
