--- app/src/main/java/com/picpay/desafio/android/presentation/view/adapter/UserListItemViewHolder.kt
@@ -1,21 +1,0 @@
-package com.picpay.desafio.android.presentation.view.adapter
-
-import androidx.recyclerview.widget.RecyclerView
-import com.bumptech.glide.Glide
-import com.picpay.desafio.android.databinding.ListItemUserBinding
-import com.picpay.desafio.android.domain.models.UsersDomain
-
-class UserListItemViewHolder(
-    private val binding: ListItemUserBinding
-) : RecyclerView.ViewHolder(binding.root) {
-
-    fun bind(user: UsersDomain) {
-        with(binding) {
-            name.text = user.name
-            username.text = user.username
-            Glide.with(root.context)
-                .load(user.img)
-                .into(picture)
-        }
-    }
-}
