+++ app/src/main/java/com/picpay/desafio/android/users/data/datasource/UsersDataSource.kt
@@ -1,0 +1,7 @@
+package com.picpay.desafio.android.users.data.datasource
+
+import com.picpay.desafio.android.users.data.models.User
+
+interface UsersDataSource {
+    suspend fun getUsers(): List<User>
+}
