+++ app/src/main/java/com/picpay/desafio/android/users/domain/models/UsersDomain.kt
@@ -1,0 +1,8 @@
+package com.picpay.desafio.android.users.domain.models
+
+data class UsersDomain(
+    val img: String?,
+    val name: String?,
+    val id: String?,
+    val username: String?
+)
