+++ app/src/main/java/com/picpay/desafio/android/users/domain/repository/UsersRepository.kt
@@ -1,0 +1,7 @@
+package com.picpay.desafio.android.users.domain.repository
+
+import com.picpay.desafio.android.users.domain.models.UsersDomain
+
+interface UsersRepository {
+    suspend fun getUsers(): List<UsersDomain>
+}
