+++ app/src/main/java/com/picpay/desafio/android/users/presentation/state/UserState.kt
@@ -1,0 +1,10 @@
+package com.picpay.desafio.android.users.presentation.state
+
+import com.picpay.desafio.android.users.domain.models.UsersDomain
+
+sealed class UserState {
+    object Inactive : UserState()
+    data class Loading(val isLoading: Boolean) : UserState()
+    data class ResponseData(val users: List<UsersDomain>?) : UserState()
+    data class Error(val error: String?) : UserState()
+}
