+++ app/src/main/java/com/picpay/desafio/android/users/presentation/usecase/UserUseCase.kt
@@ -1,0 +1,7 @@
+package com.picpay.desafio.android.users.presentation.usecase
+
+import com.picpay.desafio.android.users.domain.models.UsersDomain
+
+interface UserUseCase {
+    suspend operator fun invoke(): List<UsersDomain>
+}
