+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/colorPrimaryDark"
-    tools:context=".presentation.view.activity.MainActivity">
+    tools:context=".users.presentation.view.activity.MainActivity">
 
     <LinearLayout
         android:layout_width="match_parent"
