+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -7,7 +7,7 @@
 
     <fragment
         android:id="@+id/userListFragment"
-        android:name="com.picpay.desafio.android.presentation.view.fragment.UserListFragment"
+        android:name="com.picpay.desafio.android.users.presentation.view.fragment.UserListFragment"
         android:label="fragment_list_user"
         tools:layout="@layout/fragment_list_user" />
 </navigation>
