+++ app/src/test/java/com/picpay/desafio/android/UsersHelper.kt
@@ -1,6 +1,6 @@
 package com.picpay.desafio.android
 
-import com.picpay.desafio.android.data.models.User
+import com.picpay.desafio.android.users.data.models.User
 
 object UsersHelper {
