+++ app/src/test/java/com/picpay/desafio/android/data/datasource/UsersDataSourceImplTest.kt
@@ -1,7 +1,9 @@
 package com.picpay.desafio.android.data.datasource
 
 import com.picpay.desafio.android.UsersHelper
-import com.picpay.desafio.android.data.api.PicPayService
+import com.picpay.desafio.android.users.data.api.PicPayService
+import com.picpay.desafio.android.users.data.datasource.UsersDataSource
+import com.picpay.desafio.android.users.data.datasource.UsersDataSourceImpl
 import io.mockk.MockKAnnotations
 import io.mockk.coEvery
 import io.mockk.impl.annotations.MockK
